requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
certifi==2025.10.5
feedparser==6.0.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pydantic==2.11.10
pydantic_core==2.33.2
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.8
typing-inspection==0.4.2
typing_extensions==4.15.0
groq>=0.4.0
python-dotenv==1.0.0
markdown==3.5.1